requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
autograd==1.6.2
autograd-gamma==0.5.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.6.15.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
configparser==7.0.0
coverage==7.3.2
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.4
django-crispy-forms==2.1
django-environ==0.11.2
django-nose==1.4.7
django-shortcuts==1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
drf-spectacular==0.26.5
ecdsa==0.18.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.110.1
fastjsonschema==2.18.0
filelock==3.12.4
fitz==0.0.1.dev2
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.37.1
formulaic==0.6.4
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
html-testRunner==1.2.1
html5lib==1.1
httplib2==0.22.0
httptools==0.6.1
idna==3.4
inflection==0.5.1
interface-meta==1.3.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
ISLP==0.3.19
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
lifelines==0.27.8
lightning-utilities==0.9.0
llvmlite==0.41.1
looseversion==1.3.0
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
nose==1.3.7
notebook==7.0.2
notebook_shim==0.2.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.0.9
opt-einsum==3.3.0
oscrypto==1.3.0
outcome==1.2.0
overrides==7.4.0
package-name==0.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
pooch==1.6.0
progressbar2==4.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
prov==2.0.0
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pygam==0.9.0
Pygments==2.16.1
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.0.9
pypdf==4.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-utils==3.7.0
pytorch-lightning==2.0.9
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
rdflib==7.0.0
referencing==0.31.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
selenium==4.12.0
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.5
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
split-folders==0.5.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.2
statsmodels==0.14.0
svglib==1.5.1
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.16.0
tensorflow-intel==2.11.1
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchmetrics==1.1.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
traits==6.3.2
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.12
uvicorn==0.29.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wrapt==1.14.1
wsproto==1.2.0
xhtml2pdf==0.2.15
yarl==1.9.2
yodel==0.3.0
